test(IntegrationTestSupport): 이미지 관련 Mock 설정 추가
